[refact] signal : divide signal function
