+++ signal.c
@@ -1,0 +1,38 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   signal.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: wonyang <wonyang@student.42seoul.kr>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/01/16 17:40:34 by wonyang           #+#    #+#             */
+/*   Updated: 2023/01/16 17:42:18 by wonyang          ###   ########seoul.kr  */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include <unistd.h>
+#include <signal.h>
+#include <stdio.h>
+#include <readline/readline.h>
+#include "minishell.h"
+
+extern t_global	g_var;
+
+void	sigint_handler_prompt(int signo)
+{
+	(void)signo;
+	g_var.status = 1;
+	ft_putchar_fd('\n', STDOUT_FILENO);
+	rl_on_new_line();
+	rl_replace_line("", 0);
+	rl_redisplay();
+}
+
+void	sigint_handler_parent(int signo)
+{
+	(void)signo;
+	g_var.status = 130;
+	ft_putchar_fd('\n', STDOUT_FILENO);
+	rl_on_new_line();
+	rl_replace_line("", 0);
+}
