+++ terminal.c
@@ -6,15 +6,12 @@
 /*   By: wonyang <wonyang@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/10 15:47:32 by wonyang           #+#    #+#             */
-/*   Updated: 2023/01/16 15:47:36 by wonyang          ###   ########seoul.kr  */
+/*   Updated: 2023/01/16 17:41:45 by wonyang          ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 #include <term.h>
-#include <signal.h>
-#include <stdio.h>
-#include <readline/readline.h>
 #include "libft.h"
 #include "minishell.h"
 
@@ -27,25 +24,6 @@ static void	get_terminal_setting(void)
 	g_var.new_term.c_lflag &= ~ECHOCTL;
 	g_var.new_term.c_cc[VQUIT] = 0;
 	tcsetattr(STDIN_FILENO, TCSANOW, &(g_var.new_term));
-}
-
-void	sigint_handler_prompt(int signo)
-{
-	(void)signo;
-	g_var.status = 1;
-	ft_putchar_fd('\n', STDOUT_FILENO);
-	rl_on_new_line();
-	rl_replace_line("", 0);
-	rl_redisplay();
-}
-
-void	sigint_handler_parent(int signo)
-{
-	(void)signo;
-	g_var.status = 130;
-	ft_putchar_fd('\n', STDOUT_FILENO);
-	rl_on_new_line();
-	rl_replace_line("", 0);
 }
 
 static t_error	set_envp_value(t_envp *envp)
